requirements.txt
streamlit
mediapipe==0.10.18
opencv-python-headless
pillow
numpy
